Update fix_me_helloworld.c
The following code should run without any syntax error and output the print message-"Hello, world!".
(Additional- The \n at the end of the string adds a newline character to the output, so that the next line of text appears on a new line. In C, \n is an escape sequence that represents a newline character. It is used in print statements to indicate that the output should be moved to the beginning of the next line.)